Initilize kubernetes operator using kubebuilder
The newest not stable version 3.0.0-alpha.0 of kubebuilder
was used to generate the initial scaffolding of kubernetes operator.

The stable version 2.3.1 of kubebuilder uses older controller-gen
in v0.2.5 which creates CRDs in API version apiextensions.k8s.io/v1beta1.
This version is planed to be deprecated in kubernetes version v1.22.
The MutatingWebhookConfiguration and ValidatingWebhookConfiguration
are old too, beacuse they are in the admissionregistration.k8s.io/v1beta1
API version. The v1beta1 is planed to be deprecated in kubernetes version v1.19

Command used to generate files in commit:
kubebuilder init --domain vectorized.io --project-name redpanda-operator  \
--owner vectorized.io --plugins go.kubebuilder.io/v3-alpha